MM-112-menuoptions: move the button to hamburger folder, and menu to menu folder
